requirements.txt
Django==1.5.1
dj-database-url==0.2.1
dj-static==0.0.5
django-extensions==1.1.1
django-openid-auth==0.5
django-randomslug==0.1.3
feedparser==5.1.3
gunicorn==17.5
psycopg2==2.5.1
python-openid==2.2.5
six==1.3.0
static==0.4
wsgiref==0.1.2
